+++ lib/appsignal/capistrano.rb
@@ -1,2 +1,9 @@
 require 'appsignal'
-require 'appsignal/integrations/capistrano'
+
+if defined?(Capistrano::VERSION)
+  # Capistrano 3+
+  load File.expand_path('../integrations/capistrano/appsignal.cap', __FILE__)
+else
+  # Capistrano 2
+  require 'appsignal/integrations/capistrano/capistrano_2_tasks'
+end