+++ lib/appsignal/integrations/capistrano/capistrano_2_tasks.rb
@@ -1,5 +1,3 @@
-require 'capistrano'
-
 module Appsignal
   module Integrations
     class Capistrano
@@ -12,6 +10,7 @@ module Appsignal
             task :deploy do
               env = fetch(:rails_env, fetch(:rack_env, 'production'))
               user = ENV['USER'] || ENV['USERNAME']
+              revision = fetch(:appsignal_revision, fetch(:current_revision))
 
               appsignal_config = Appsignal::Config.new(
                 ENV['PWD'],
@@ -22,8 +21,7 @@ module Appsignal
 
               if appsignal_config && appsignal_config.active?
                 marker_data = {
-                  :revision => current_revision,
-                  :repository => repository,
+                  :revision => revision,
                   :user => user
                 }