+++ lib/appsignal/integrations/delayed_job.rb
@@ -26,12 +26,10 @@ if defined?(::Delayed::Plugin)
               block.call(job)
             end
           rescue Exception => exception
-            unless Appsignal.is_ignored_exception?(exception)
-              Appsignal::Transaction.current.add_exception(exception)
-            end
+            Appsignal.add_exception(exception)
             raise exception
           ensure
-            Appsignal::Transaction.current.complete!
+            Appsignal::Transaction.complete_current!
           end
         end
       end