+++ lib/appsignal/integrations/sidekiq.rb
@@ -17,12 +17,10 @@ if defined?(::Sidekiq)
             yield
           end
         rescue Exception => exception
-          unless Appsignal.is_ignored_exception?(exception)
-            Appsignal::Transaction.current.add_exception(exception)
-          end
+          Appsignal.add_exception(exception)
           raise exception
         ensure
-          Appsignal::Transaction.current.complete!
+          Appsignal::Transaction.complete_current!
         end
       end
     end