+++ lib/appsignal/rack/listener.rb
@@ -2,6 +2,7 @@ module Appsignal
   module Rack
     class Listener
       def initialize(app, options = {})
+        Appsignal.logger.debug 'Initializing Appsignal::Rack::Listener'
         @app, @options = app, options
       end
 
@@ -17,12 +18,10 @@ module Appsignal
         Appsignal::Transaction.create(request_id(env), env)
         @app.call(env)
       rescue Exception => exception
-        unless Appsignal.is_ignored_exception?(exception)
-          Appsignal::Transaction.current.add_exception(exception)
-        end
+        Appsignal.add_exception(exception)
         raise exception
       ensure
-        Appsignal::Transaction.current.complete!
+        Appsignal::Transaction.complete_current!
       end
 
       def request_id(env)