+++ lib/appsignal/transaction.rb
@@ -11,20 +11,30 @@
     HTTP_CACHE_CONTROL HTTP_CONNECTION HTTP_USER_AGENT HTTP_FROM HTTP_NEGOTIATE
     HTTP_PRAGMA HTTP_REFERER HTTP_X_FORWARDED_FOR).freeze
 
-    def self.create(key, env)
-      Appsignal.logger.debug("Creating transaction: #{key}")
-      Thread.current[:appsignal_transaction_id] = key
-      Appsignal.transactions[key] = Appsignal::Transaction.new(key, env)
+    def self.create(request_id, env)
+      Appsignal.logger.debug("Creating transaction: #{request_id}")
+      Thread.current[:appsignal_transaction_id] = request_id
+      Appsignal::Transaction.new(request_id, env)
     end
 
     def self.current
       Appsignal.transactions[Thread.current[:appsignal_transaction_id]]
     end
 
+    def self.complete_current!
+      if current
+        current.complete!
+        Thread.current[:appsignal_transaction_id] = nil
+      else
+        Appsignal.logger.error('Trying to complete current, but no transaction present')
+      end
+    end
+
     attr_reader :request_id, :events, :process_action_event, :action, :exception,
                 :env, :fullpath, :time, :tags, :kind, :queue_start
 
     def initialize(request_id, env)
+      Appsignal.transactions[request_id] = self
       @request_id = request_id
       @events = []
       @process_action_event = nil
@@ -116,17 +126,17 @@
 
     def complete!
       Appsignal.logger.debug("Completing transaction: #{@request_id}")
-      Thread.current[:appsignal_transaction_id] = nil
-      current_transaction = Appsignal.transactions.delete(@request_id)
       if process_action_event || exception?
         if Appsignal::Pipe.current
           Appsignal::Pipe.current.write(self)
         else
-          Appsignal.enqueue(current_transaction)
+          Appsignal.enqueue(self)
         end
       else
         Appsignal.logger.debug("No process_action_event or exception: #{@request_id}")
       end
+    ensure
+      Appsignal.transactions.delete(@request_id)
     end
 
     def set_background_queue_start