+++ spec/lib/appsignal/marker_spec.rb
@@ -1,5 +1,4 @@
 require 'spec_helper'
-require 'capistrano'
 
 describe Appsignal::Marker do
   let(:config) { project_fixture_config }
@@ -44,7 +43,7 @@ describe Appsignal::Marker do
 
           marker.transmit
 
-          log.string.should include('Notifying Appsignal of deploy...')
+          log.string.should include('Notifying Appsignal of deploy with: revision: 503ce0923ed177a3ce000005, user: batman')
           log.string.should include('Appsignal has been notified of this deploy!')
         end
 
@@ -54,7 +53,7 @@ describe Appsignal::Marker do
 
           marker.transmit
 
-          log.string.should include('Notifying Appsignal of deploy...')
+          log.string.should include('Notifying Appsignal of deploy with: revision: 503ce0923ed177a3ce000005, user: batman')
           log.string.should include(
             'Something went wrong while trying to notify Appsignal: 500 at http://localhost:3000/1/markers'
           )
@@ -66,14 +65,18 @@ describe Appsignal::Marker do
 
       it_should_behave_like "logging info and errors"
 
-      context "with a Capistrano logger" do
-        let(:logger) {
-          Capistrano::Logger.new(:output => log).tap do |logger|
-            logger.level = Capistrano::Logger::MAX_LEVEL
-          end
-        }
+      if capistrano2_present?
+        require 'capistrano'
 
-        it_should_behave_like "logging info and errors"
+        context "with a Capistrano 2 logger" do
+          let(:logger) {
+            Capistrano::Logger.new(:output => log).tap do |logger|
+              logger.level = Capistrano::Logger::MAX_LEVEL
+            end
+          }
+
+          it_should_behave_like "logging info and errors"
+        end
       end
     end
   end